requirements.txt
Pillow==9.4.0
transformers==4.26.1
